+++ src/app/api/toko/sales/route.ts
@@ -76,9 +76,9 @@ export async function POST(request: Request) {
             return NextResponse.json({ message: "Keranjang kosong" }, { status: 400 });
         }
 
-        // Server-side validation: reject negative/zero quantities
+        // Server-side validation: reject negative/zero quantities (parseFloat untuk desimal)
         for (const item of items) {
-            const qty = parseInt(item.quantity);
+            const qty = parseFloat(item.quantity);
             if (!qty || qty <= 0 || isNaN(qty)) {
                 return NextResponse.json({ message: "Jumlah item harus lebih dari 0" }, { status: 400 });
             }
